fix(tests): stub dispatch metadata suite-wide for no-dist runs
The dispatch-metadata hermeticity stub lived in tests/unit/docker/conftest.py,
so it only reached docker/ tests. Baseline-container and engine docker-path
tests (tests/unit/study/test_baseline_container.py,
tests/unit/engines/test_vllm_engine.py) also real-execute build_docker_cmd ->
append_package_dispatch -> importlib.metadata.requires and failed with
PackageNotFoundError when the suite runs in-container from a source tree on
PYTHONPATH with no install.

Hoist the autouse stub_runtime_requirements fixture to tests/unit/conftest.py
so it covers the whole unit suite. The two needs_dist_metadata-marked tests are
unchanged and keep working (the marker logic travels with the fixture: skipped
when the distribution is absent, run for real when installed).
tests/unit/docker/conftest.py keeps make_subprocess_result.
